Updated few user, posts related logic in all classes in this chatbook app

+++ src/main/java/in/krish/binding/CreatePost.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class CreatePost {
 	
-	private Integer postid;
+	private Long postid;
 	private String title;
 	
 	private String description;
